improvements in the control of alt currencies when launching the app
